fixes + handling of multiple networks per layer

+++ setup_cpu.lua
@@ -1,6 +1,6 @@
 require('nn')
 
-function transfer_data(m) return m:double() end
+function transfer_data(m) return m:float() end
 
 local function make_deterministic(seed)
   torch.manualSeed(seed)
